chore(countdown): demo拆解与规范
