+++ src/packages/countdown/demos/h5/demo5.tsx
@@ -1,0 +1,18 @@
+import React, { useRef } from 'react'
+import { Cell, CountDown } from '@nutui/nutui-react'
+
+const Demo5 = () => {
+  const stateRef = useRef({
+    serverTime: Date.now() - 20 * 1000,
+    endTime: Date.now() + 60 * 1000,
+  })
+  return (
+    <Cell>
+      <CountDown
+        startTime={stateRef.current.serverTime}
+        endTime={stateRef.current.endTime}
+      />
+    </Cell>
+  )
+}
+export default Demo5
