+++ src/packages/countdown/demos/h5/demo6.tsx
@@ -1,0 +1,24 @@
+import React, { useEffect, useRef, useState } from 'react'
+import { Cell, CountDown } from '@nutui/nutui-react'
+
+const Demo6 = () => {
+  const [asyncEnd, setAsyncEnd] = useState(0)
+  const stateRef = useRef({
+    timer: -1,
+    endTime: Date.now() + 60 * 1000,
+  })
+  useEffect(() => {
+    stateRef.current.timer = window.setTimeout(() => {
+      setAsyncEnd(Date.now() + 30 * 1000)
+    }, 3000)
+    return () => {
+      clearTimeout(stateRef.current.timer)
+    }
+  }, [])
+  return (
+    <Cell>
+      <CountDown endTime={asyncEnd} />
+    </Cell>
+  )
+}
+export default Demo6
