+++ src/packages/countdown/doc.md
@@ -16,25 +16,7 @@ import { CountDown } from '@nutui/nutui-react'
 
 :::demo
 
-```tsx
-import  React, {useRef }from "react";
-import { Cell, CountDown } from '@nutui/nutui-react';
-
-const App = () => {
-  const stateRef = useRef({
-    endTime: Date.now() + 60 * 1000,
-  })
-  const onEnd = () => {
-    console.log('countdown: ended.')
-  }
-  return (
-     <Cell>
-        <CountDown endTime={stateRef.current.endTime} onEnd={onEnd} />
-    </Cell>
-  );
-};
-export default App;
-```
+<CodeBlock src='h5/demo1.tsx'></CodeBlock>
 
 :::
 
@@ -42,22 +24,7 @@ export default App;
 
 :::demo
 
-```tsx
-import  React, {useRef }from "react";
-import { Cell, CountDown } from '@nutui/nutui-react';
-
-const App = () => {
-  const stateRef = useRef({
-    remainingTime:  60 * 1000,
-  })
-  return (
-     <Cell>
-         <CountDown remainingTime={stateRef.current.remainingTime} />
-    </Cell>
-  );
-};
-export default App;
-```
+<CodeBlock src='h5/demo2.tsx'></CodeBlock>
 
 :::
 
@@ -65,22 +32,7 @@ export default App;
 
 :::demo
 
-```tsx
-import  React, {useRef }from "react";
-import { Cell, CountDown } from '@nutui/nutui-react';
-
-const App = () => {
-  const stateRef = useRef({
-    endTime: Date.now() + 60 * 1000,
-  })
-  return (
-     <Cell>
-        <CountDown endTime={stateRef.current.endTime} format="DD 天 HH 时 mm 分 ss 秒"/>
-    </Cell>
-  );
-};
-export default App;
-```
+<CodeBlock src='h5/demo3.tsx'></CodeBlock>
 
 :::
 
@@ -88,23 +40,7 @@ export default App;
 
 :::demo
 
-```tsx
-import  React, {useRef }from "react";
-import { Cell, CountDown } from '@nutui/nutui-react';
-
-const App = () => {
-  const stateRef = useRef({
-    endTime: Date.now() + 60 * 1000,
-  })
-  return (
-     <Cell>
-        <CountDown endTime={stateRef.current.endTime} millisecond format="HH:mm:ss:SS"
-        />
-    </Cell>
-  );
-};
-export default App;
-```
+<CodeBlock src='h5/demo4.tsx'></CodeBlock>
 
 :::
 
@@ -112,23 +48,7 @@ export default App;
 
 :::demo
 
-```tsx
-import  React, {useRef }from "react";
-import { Cell, CountDown } from '@nutui/nutui-react';
-
-const App = () => {
-  const stateRef = useRef({
-    serverTime: Date.now() - 20 * 1000,
-    endTime: Date.now() + 60 * 1000,
-  })
-  return (
-    <Cell>
-        <CountDown startTime={stateRef.current.serverTime} endTime={stateRef.current.endTime} />
-    </Cell>
-  );
-};
-export default App;
-```
+<CodeBlock src='h5/demo5.tsx'></CodeBlock>
 
 :::
 
@@ -136,32 +56,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React, { useEffect, useRef, useState } from 'react'
-import { Cell, CountDown } from '@nutui/nutui-react';
-
-const App = () => {
-  const [asyncEnd, setAsyncEnd] = useState(0)
-  const stateRef = useRef({
-    timer: -1,
-    endTime: Date.now() + 60 * 1000,
-  })
-  useEffect(() => {
-    stateRef.current.timer = setTimeout(() => {
-      setAsyncEnd(Date.now() + 30 * 1000)
-    }, 3000)
-    return () => {
-      clearTimeout(stateRef.current.timer)
-    }
-  }, [])
-  return (
-    <Cell>
-        <CountDown  endTime={asyncEnd} />
-    </Cell>
-  );
-};
-export default App;
-```
+<CodeBlock src='h5/demo6.tsx'></CodeBlock>
 
 :::
 
@@ -169,45 +64,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React, {  useRef, useState } from 'react'
-import { Cell, CountDown, Button } from '@nutui/nutui-react';
-
-const App = () => {
-  const stateRef = useRef({
-    endTime: Date.now() + 60 * 1000,
-  })
-  const [paused, setPaused] = useState(false)
-  const toggle = () => {
-    console.log(paused)
-    setPaused(!paused)
-  }
-  const onpaused = (v) => {
-    console.log('paused: ', v)
-  }
-  const onrestart = (v) => {
-    console.log('restart: ', v)
-  }
-  return (
-    <Cell>
-        <CountDown
-            endTime={stateRef.current.endTime}
-            paused={paused}
-            onPaused={onpaused}
-            onRestart={onrestart}
-          />
-          <div style={{ position: 'absolute', right: '10px', top: '9px' }}>
-            <div onClick={() => toggle()}>
-              <Button type="primary" size="small">
-                {paused ? 'start' : 'stop'}
-              </Button>
-            </div>
-          </div>
-    </Cell>
-  );
-};
-export default App;
-```
+<CodeBlock src='h5/demo7.tsx'></CodeBlock>
 
 :::
 
@@ -215,71 +72,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React, {  useRef, useState } from 'react'
-import { Cell, CountDown } from '@nutui/nutui-react';
-
-
-const partItemStyle = {
-    flexShrink: 0,
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    width: '20px',
-    height: '25px',
-    background: '#e8220e',
-    color: '#fff',
-    fontSize: '14px',
-    borderRadius: '6px',
-}
-const partItemSymbolStyle = {
-    margin: '0 5px',
-}
-const App = () => {
-  const onUpdate = (v) => {
-    setResetTime(v)
-  }
-  const [resetTime, setResetTime] = useState({
-    d: '1',
-    h: '00',
-    m: '00',
-    s: '00',
-  })
-  const stateRef = useRef({
-    endTime: Date.now() + 60 * 1000,
-  })
-  
-  return (
-    <Cell>
-        <span>
-        <CountDown endTime={stateRef.current.endTime} onUpdate={onUpdate}>
-            <div className="countdown-part-box" style={{ display: 'flex', alignItems: 'center' }}>
-            <div className="part-item-symbol" style={partItemSymbolStyle}>
-                {resetTime.d}天
-            </div>
-            <div className="part-item h" style={partItemStyle}>
-                {resetTime.h}
-            </div>
-            <span className="part-item-symbol" style={partItemSymbolStyle}>
-                :
-            </span>
-            <div className="part-item m" style={partItemStyle}>
-                {resetTime.m}
-            </div>
-            <span className="part-item-symbol" style={partItemSymbolStyle}>
-                :
-            </span>
-            <div className="part-item s" style={partItemStyle}>
-                {resetTime.s}
-            </div>
-            </div>
-        </CountDown>
-        </span>
-    </Cell>
-  );
-};
-export default App;
-```
+<CodeBlock src='h5/demo8.tsx'></CodeBlock>
 
 :::
 
@@ -289,57 +82,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React, {  useRef } from 'react'
-import { Cell, CountDown, Grid, GridItem, Button } from '@nutui/nutui-react';
-
-const App = () => {
-  
-  const countDownRef = useRef<countdownRefState>(null)
-  const start = () => {
-    console.log(countDownRef.current)
-    countDownRef.current && countDownRef.current.start()
-  }
-
-  const pause = () => {
-    countDownRef.current && countDownRef.current.pause()
-  }
-
-  const reset = () => {
-    countDownRef.current && countDownRef.current.reset()
-  }
-  return (
-    <>
-    <Cell>
-          <CountDown
-            format="ss:SS"
-            autoStart={false}
-            time={20000}
-            ref={countDownRef}
-          />
-        </Cell>
-        <Grid columns="3">
-          <GridItem>
-            <Button type="primary" onClick={start}>
-              开始
-            </Button>
-          </GridItem>
-          <GridItem>
-            <Button type="primary" onClick={pause}>
-              暂停
-            </Button>
-          </GridItem>
-          <GridItem>
-            <Button type="primary" onClick={reset}>
-              重置
-            </Button>
-          </GridItem>
-        </Grid>
-    </>
-  );
-};
-export default App;
-```
+<CodeBlock src='h5/demo9.tsx'></CodeBlock>
 
 :::
